CFIN-402 Rename tags to be rich tags
