+++ src/tests/components/RichTag.test.js
@@ -1,10 +1,10 @@
 import { render, screen, within } from "@testing-library/react";
-import { Tag } from "../../components/Tag";
+import { RichTag } from "../../components/RichTag";
 import React from "react";
 
 describe("Tag", () => {
     it("displays tag containing given icon and text", () => {
-        render(<Tag topIcon="battery-full" title="Battery is full" />);
+        render(<RichTag topIcon="battery-full" title="Battery is full" />);
 
         const tag = screen.getByRole("listitem");
         expect(tag).toHaveTextContent("Battery is full");
